set uniform file structure across nfactpp. Changed import statements
